refactor: api-service 컨트롤러 분리
