+++ auth-service/src/main/java/com/sobok/authservice/auth/dto/request/AuthResetPwReqDto.java
@@ -13,6 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AuthResetPwReqDto {
     private Long authId;
+    private String inputCode;
 
     @NotBlank(message = "비밀번호는 필수 입니다.")
     @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}",
